requirements.txt
git+https://github.com/Havilash/Neural-Network.git@main#egg=neural_network
flask
numpy
pillow
